requirements.txt
pyserial
kivy